refactor: reorganize code files
