+++ plugins/manager.go
@@ -10,30 +10,18 @@ package plugins
 
 import (
 	"context"
-	"fmt"
-	"maps"
 	"os"
-	"path/filepath"
 	"sync"
-	"time"
 
 	"github.com/navidrome/navidrome/conf"
 	"github.com/navidrome/navidrome/core/agents"
 	"github.com/navidrome/navidrome/core/scrobbler"
 	"github.com/navidrome/navidrome/log"
 	"github.com/navidrome/navidrome/plugins/api"
-	"github.com/navidrome/navidrome/plugins/host/artwork"
-	"github.com/navidrome/navidrome/plugins/host/cache"
-	"github.com/navidrome/navidrome/plugins/host/config"
-	"github.com/navidrome/navidrome/plugins/host/http"
-	"github.com/navidrome/navidrome/plugins/host/scheduler"
-	"github.com/navidrome/navidrome/plugins/host/websocket"
 	"github.com/navidrome/navidrome/plugins/schema"
 	"github.com/navidrome/navidrome/utils/singleton"
 	"github.com/navidrome/navidrome/utils/slice"
 	"github.com/tetratelabs/wazero"
-	wazeroapi "github.com/tetratelabs/wazero/api"
-	"github.com/tetratelabs/wazero/imports/wasi_snapshot_preview1"
 )
 
 const (
@@ -60,45 +48,6 @@ type WasmPlugin interface {
 	PluginID() string
 	// Instantiate creates a new instance of the plugin and returns it along with a cleanup function
 	Instantiate(ctx context.Context) (any, func(), error)
-}
-
-const maxParallelCompilations = 2 // Limit to 2 concurrent compilations
-
-var (
-	compileSemaphore = make(chan struct{}, maxParallelCompilations)
-	compilationCache wazero.CompilationCache
-	cacheOnce        sync.Once
-	runtimePool      sync.Map // map[string]*pooledRuntime
-)
-
-func getCompilationCache() (wazero.CompilationCache, error) {
-	var err error
-	cacheOnce.Do(func() {
-		cacheDir := filepath.Join(conf.Server.CacheFolder, "plugins")
-		purgeCacheBySize(cacheDir, conf.Server.Plugins.CacheSize)
-		compilationCache, err = wazero.NewCompilationCacheWithDir(cacheDir)
-	})
-	return compilationCache, err
-}
-
-type pluginState struct {
-	ready chan struct{}
-	err   error
-}
-
-// Helper to create the correct ModuleConfig for plugins
-func newWazeroModuleConfig() wazero.ModuleConfig {
-	return wazero.NewModuleConfig().WithStartFunctions("_initialize").WithStderr(log.Writer())
-}
-
-// PluginDiscoveryEntry represents the result of plugin discovery
-type PluginDiscoveryEntry struct {
-	ID        string                 // Plugin ID (directory name)
-	Path      string                 // Resolved plugin directory path
-	WasmPath  string                 // Path to the WASM file
-	Manifest  *schema.PluginManifest // Loaded manifest (nil if failed)
-	IsSymlink bool                   // Whether the plugin is a development symlink
-	Error     error                  // Error encountered during discovery
 }
 
 type pluginInfo struct {
@@ -141,191 +90,6 @@ func createManager() *Manager {
 	m.websocketService = newWebsocketService(m)
 
 	return m
-}
-
-// precompilePlugin compiles the WASM module in the background and updates the pluginState.
-func precompilePlugin(state *pluginState, customRuntime api.WazeroNewRuntime, wasmPath, name string) {
-	compileSemaphore <- struct{}{}
-	defer func() { <-compileSemaphore }()
-	ctx := context.Background()
-	r, err := customRuntime(ctx)
-	if err != nil {
-		state.err = fmt.Errorf("failed to create runtime for plugin %s: %w", name, err)
-		close(state.ready)
-		return
-	}
-	defer r.Close(ctx)
-	b, err := os.ReadFile(wasmPath)
-	if err != nil {
-		state.err = fmt.Errorf("failed to read wasm file: %w", err)
-		close(state.ready)
-		return
-	}
-	if _, err := r.CompileModule(ctx, b); err != nil {
-		state.err = fmt.Errorf("failed to compile WASM for plugin %s: %w", name, err)
-		log.Warn("Plugin compilation failed", "name", name, "path", wasmPath, "err", err)
-	} else {
-		state.err = nil
-		log.Debug("Plugin compilation completed", "name", name, "path", wasmPath)
-	}
-	close(state.ready)
-}
-
-func pluginCompilationTimeout() time.Duration {
-	if conf.Server.DevPluginCompilationTimeout > 0 {
-		return conf.Server.DevPluginCompilationTimeout
-	}
-	return time.Minute
-}
-
-// waitForPluginReady blocks until the plugin is compiled and returns true if ready, false otherwise.
-func waitForPluginReady(state *pluginState, pluginID, wasmPath string) bool {
-	timeout := pluginCompilationTimeout()
-	select {
-	case <-state.ready:
-	case <-time.After(timeout):
-		log.Error("Timed out waiting for plugin compilation", "name", pluginID, "path", wasmPath, "timeout", timeout)
-		return false
-	}
-	if state.err != nil {
-		log.Error("Failed to compile plugin", "name", pluginID, "path", wasmPath, state.err)
-		return false
-	}
-	return true
-}
-
-// loadHostLibrary loads the given host library and returns its exported functions
-func loadHostLibrary[S any](
-	ctx context.Context,
-	instantiateFn func(context.Context, wazero.Runtime, S) error,
-	service S,
-) (map[string]wazeroapi.FunctionDefinition, error) {
-	r := wazero.NewRuntime(ctx)
-	if err := instantiateFn(ctx, r, service); err != nil {
-		return nil, err
-	}
-	m := r.Module("env")
-	return m.ExportedFunctionDefinitions(), nil
-}
-
-// combineLibraries combines the given host libraries into a single "env" module
-func (m *Manager) combineLibraries(ctx context.Context, r wazero.Runtime, libs ...map[string]wazeroapi.FunctionDefinition) error {
-	// Merge the libraries
-	hostLib := map[string]wazeroapi.FunctionDefinition{}
-	for _, lib := range libs {
-		maps.Copy(hostLib, lib)
-	}
-
-	// Create the combined host module
-	envBuilder := r.NewHostModuleBuilder("env")
-	for name, fd := range hostLib {
-		fn, ok := fd.GoFunction().(wazeroapi.GoModuleFunction)
-		if !ok {
-			return fmt.Errorf("invalid function definition: %s", fd.DebugName())
-		}
-		envBuilder.NewFunctionBuilder().
-			WithGoModuleFunction(fn, fd.ParamTypes(), fd.ResultTypes()).
-			WithParameterNames(fd.ParamNames()...).Export(name)
-	}
-
-	// Instantiate the combined host module
-	if _, err := envBuilder.Instantiate(ctx); err != nil {
-		return err
-	}
-	return nil
-}
-
-// createCustomRuntime returns a function that creates a new wazero runtime with the given compilation cache
-// and instantiates the required host functions
-func (m *Manager) createCustomRuntime(compCache wazero.CompilationCache, pluginID string, permissions schema.PluginManifestPermissions) api.WazeroNewRuntime {
-	return func(ctx context.Context) (wazero.Runtime, error) {
-		// Check if runtime already exists
-		if rt, ok := runtimePool.Load(pluginID); ok {
-			log.Trace(ctx, "Using existing runtime", "plugin", pluginID, "runtime", fmt.Sprintf("%p", rt))
-			return rt.(wazero.Runtime), nil
-		}
-
-		// Create the runtime
-		runtimeConfig := wazero.NewRuntimeConfig().WithCompilationCache(compCache)
-		r := wazero.NewRuntimeWithConfig(ctx, runtimeConfig)
-		if _, err := wasi_snapshot_preview1.Instantiate(ctx, r); err != nil {
-			return nil, err
-		}
-
-		// Define all available host services
-		type hostService struct {
-			name        string
-			isPermitted bool
-			loadFunc    func() (map[string]wazeroapi.FunctionDefinition, error)
-		}
-
-		// List of all available host services with their permissions and loading functions. For each service, we check
-		// if the plugin has the required permission before loading it.
-		availableServices := []hostService{
-			{"config", permissions.Config != nil, func() (map[string]wazeroapi.FunctionDefinition, error) {
-				return loadHostLibrary[config.ConfigService](ctx, config.Instantiate, &configServiceImpl{pluginID: pluginID})
-			}},
-			{"scheduler", permissions.Scheduler != nil, func() (map[string]wazeroapi.FunctionDefinition, error) {
-				return loadHostLibrary[scheduler.SchedulerService](ctx, scheduler.Instantiate, m.schedulerService.HostFunctions(pluginID))
-			}},
-			{"cache", permissions.Cache != nil, func() (map[string]wazeroapi.FunctionDefinition, error) {
-				return loadHostLibrary[cache.CacheService](ctx, cache.Instantiate, newCacheService(pluginID))
-			}},
-			{"artwork", permissions.Artwork != nil, func() (map[string]wazeroapi.FunctionDefinition, error) {
-				return loadHostLibrary[artwork.ArtworkService](ctx, artwork.Instantiate, &artworkServiceImpl{})
-			}},
-			{"http", permissions.Http != nil, func() (map[string]wazeroapi.FunctionDefinition, error) {
-				httpPerms, err := parseHTTPPermissions(permissions.Http)
-				if err != nil {
-					return nil, fmt.Errorf("invalid http permissions for plugin %s: %w", pluginID, err)
-				}
-				return loadHostLibrary[http.HttpService](ctx, http.Instantiate, &httpServiceImpl{
-					pluginID:    pluginID,
-					permissions: httpPerms,
-				})
-			}},
-			{"websocket", permissions.Websocket != nil, func() (map[string]wazeroapi.FunctionDefinition, error) {
-				wsPerms, err := parseWebSocketPermissions(permissions.Websocket)
-				if err != nil {
-					return nil, fmt.Errorf("invalid websocket permissions for plugin %s: %w", pluginID, err)
-				}
-				return loadHostLibrary[websocket.WebSocketService](ctx, websocket.Instantiate, m.websocketService.HostFunctions(pluginID, wsPerms))
-			}},
-		}
-
-		// Load only permitted services
-		var grantedPermissions []string
-		var libraries []map[string]wazeroapi.FunctionDefinition
-		for _, service := range availableServices {
-			if service.isPermitted {
-				lib, err := service.loadFunc()
-				if err != nil {
-					return nil, fmt.Errorf("error loading %s lib: %w", service.name, err)
-				}
-				libraries = append(libraries, lib)
-				grantedPermissions = append(grantedPermissions, service.name)
-			}
-		}
-		log.Trace(ctx, "Granting permissions for plugin", "plugin", pluginID, "permissions", grantedPermissions)
-
-		// Combine the permitted libraries
-		if err := m.combineLibraries(ctx, r, libraries...); err != nil {
-			return nil, err
-		}
-
-		pooled := newPooledRuntime(r, pluginID)
-
-		// Use LoadOrStore to atomically check and store, preventing race conditions
-		if existing, loaded := runtimePool.LoadOrStore(pluginID, pooled); loaded {
-			// Another goroutine created the runtime first, close ours and return the existing one
-			log.Trace(ctx, "Race condition detected, using existing runtime", "plugin", pluginID, "runtime", fmt.Sprintf("%p", existing))
-			_ = r.Close(ctx)
-			return existing.(wazero.Runtime), nil
-		}
-		log.Trace(ctx, "Created new runtime", "plugin", pluginID, "runtime", fmt.Sprintf("%p", pooled))
-
-		return pooled, nil
-	}
 }
 
 // registerPlugin adds a plugin to the registry with the given parameters
@@ -403,132 +167,6 @@ func (m *Manager) initializePluginIfNeeded(plugin *pluginInfo) {
 			break
 		}
 	}
-}
-
-// DiscoverPlugins scans the plugins directory and returns information about all discoverable plugins
-// This shared function eliminates duplication between ScanPlugins and plugin list commands
-func DiscoverPlugins(pluginsDir string) []PluginDiscoveryEntry {
-	var discoveries []PluginDiscoveryEntry
-
-	entries, err := os.ReadDir(pluginsDir)
-	if err != nil {
-		// Return a single entry with the error
-		return []PluginDiscoveryEntry{{
-			Error: fmt.Errorf("failed to read plugins directory %s: %w", pluginsDir, err),
-		}}
-	}
-
-	for _, entry := range entries {
-		name := entry.Name()
-		pluginPath := filepath.Join(pluginsDir, name)
-
-		// Skip hidden files
-		if name[0] == '.' {
-			continue
-		}
-
-		// Check if it's a directory or symlink
-		info, err := os.Lstat(pluginPath)
-		if err != nil {
-			discoveries = append(discoveries, PluginDiscoveryEntry{
-				ID:    name,
-				Error: fmt.Errorf("failed to stat entry %s: %w", pluginPath, err),
-			})
-			continue
-		}
-
-		isSymlink := info.Mode()&os.ModeSymlink != 0
-		isDir := info.IsDir()
-
-		// Skip if not a directory or symlink
-		if !isDir && !isSymlink {
-			continue
-		}
-
-		// Resolve symlinks
-		pluginDir := pluginPath
-		if isSymlink {
-			targetDir, err := os.Readlink(pluginPath)
-			if err != nil {
-				discoveries = append(discoveries, PluginDiscoveryEntry{
-					ID:        name,
-					IsSymlink: true,
-					Error:     fmt.Errorf("failed to resolve symlink %s: %w", pluginPath, err),
-				})
-				continue
-			}
-
-			// If target is a relative path, make it absolute
-			if !filepath.IsAbs(targetDir) {
-				targetDir = filepath.Join(filepath.Dir(pluginPath), targetDir)
-			}
-
-			// Verify that the target is a directory
-			targetInfo, err := os.Stat(targetDir)
-			if err != nil {
-				discoveries = append(discoveries, PluginDiscoveryEntry{
-					ID:        name,
-					IsSymlink: true,
-					Error:     fmt.Errorf("failed to stat symlink target %s: %w", targetDir, err),
-				})
-				continue
-			}
-
-			if !targetInfo.IsDir() {
-				discoveries = append(discoveries, PluginDiscoveryEntry{
-					ID:        name,
-					IsSymlink: true,
-					Error:     fmt.Errorf("symlink target is not a directory: %s", targetDir),
-				})
-				continue
-			}
-
-			pluginDir = targetDir
-		}
-
-		// Check for WASM file
-		wasmPath := filepath.Join(pluginDir, "plugin.wasm")
-		if _, err := os.Stat(wasmPath); err != nil {
-			discoveries = append(discoveries, PluginDiscoveryEntry{
-				ID:    name,
-				Path:  pluginDir,
-				Error: fmt.Errorf("no plugin.wasm found: %w", err),
-			})
-			continue
-		}
-
-		// Load manifest
-		manifest, err := LoadManifest(pluginDir)
-		if err != nil {
-			discoveries = append(discoveries, PluginDiscoveryEntry{
-				ID:    name,
-				Path:  pluginDir,
-				Error: fmt.Errorf("failed to load manifest: %w", err),
-			})
-			continue
-		}
-
-		// Check for capabilities
-		if len(manifest.Capabilities) == 0 {
-			discoveries = append(discoveries, PluginDiscoveryEntry{
-				ID:    name,
-				Path:  pluginDir,
-				Error: fmt.Errorf("no capabilities found in manifest"),
-			})
-			continue
-		}
-
-		// Success!
-		discoveries = append(discoveries, PluginDiscoveryEntry{
-			ID:        name,
-			Path:      pluginDir,
-			WasmPath:  wasmPath,
-			Manifest:  manifest,
-			IsSymlink: isSymlink,
-		})
-	}
-
-	return discoveries
 }
 
 // ScanPlugins scans the plugins directory and compiles all valid plugins without registering them.
